+++ src/preload.ts
@@ -26,7 +26,6 @@ contextBridge.exposeInMainWorld('mediaAPI', {
   showCustomPresetFile: (): Promise<boolean> => ipcRenderer.invoke('custom-presets:show'),
   releasePreviews: (ids: string[]): Promise<void> => ipcRenderer.invoke('source:release-previews', ids),
   minimizeWindow: (): Promise<void> => ipcRenderer.invoke('window:minimize'),
-  invalidateWindow: (): void => ipcRenderer.send('window:invalidate'),
   closeWindow: (): Promise<void> => ipcRenderer.invoke('window:close'),
   finishAndClose: (): Promise<void> => ipcRenderer.invoke('app:finish-and-close'),
   checkForUpdates: (): Promise<string> => ipcRenderer.invoke('app:check-update'),
@@ -78,7 +77,6 @@ declare global {
       showCustomPresetFile: () => Promise<boolean>;
       releasePreviews: (ids: string[]) => Promise<void>;
       minimizeWindow: () => Promise<void>;
-      invalidateWindow: () => void;
       closeWindow: () => Promise<void>;
       finishAndClose: () => Promise<void>;
       checkForUpdates: () => Promise<string>;